Refactor screens folders names
