+++ lib/ui/screens/add_sight/add_sight_screen_helper.dart
