+++ lib/ui/screens/sight_search/sight_search_screen_helper.dart
